Fix reference to new event builder after renaming.
