+++ lib/core/event_builder/tests.js
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-var eventBuilder = require('./index_v3.js');
+var eventBuilder = require('./index.js');
 var packageJSON = require('../../../package.json');
 var projectConfig = require('../project_config');
 var testData = require('../../../tests/test_data');
