+++ lib/optimizely/tests.js
@@ -18,7 +18,7 @@ var Optimizely = require('./');
 var bluebird = require('bluebird');
 var bucketer = require('../core/bucketer');
 var enums = require('../utils/enums');
-var eventBuilder = require('../core/event_builder/index_v3.js');
+var eventBuilder = require('../core/event_builder/index.js');
 var eventDispatcher = require('../plugins/event_dispatcher');
 var errorHandler = require('../plugins/error_handler');
 var jsonSchemaValidator = require('../utils/json_schema_validator');
